refactor(stella-cli): split run_turn and the output envelope out of two god files
crates/stella-cli/src/agent.rs and src/main.rs both sat within a handful
of lines of the 1500-line file-size ratchet, with neither grandfathered,
so the next feature touching either would be forced into an unplanned
split mid-change.

Pure move, no behavior change:

- agent.rs: run_turn and its private helper mirror_task_board move to
  agent/turn.rs (agent.rs already sits beside an agent/ directory, so
  this is the foo.rs-beside-foo/ layout the workspace uses everywhere).
  `use super::*` carries the parent's scope over; agent.rs re-exports
  `turn::{run_turn, mirror_task_board}` so every caller keeps its path,
  the same shape cli.rs -> cli/command.rs used (#3776). mirror_task_board
  widens from private to pub(crate): its only out-of-module caller,
  agent::tests, is now a sibling rather than a descendant.
- main.rs: the --output-format json|stream-json envelope contract
  (OutputFormat, SUMMARY_SCHEMA_VERSION, the pre-flight error envelope,
  emit_error_summary) moves to output_envelope.rs, a fresh sibling
  module. main.rs re-exports it the same way it already re-exports
  cli::Command, so crate::OutputFormat and friends are unchanged.
  emit_error_summary widens from private to pub(crate) for the same
  reason mirror_task_board does: its only caller, main(), stays behind.

check-file-size.sh passes with no baseline entries added or changed for
either file.

Two of the moved lines needed a prose-baseline hand-edit, following the
precedent #5422 set for the identical shape: `scripts/prose-baseline.txt`
and `scripts/prose-grade-baseline.txt` are keyed per file, and `--update`
carries an entry to a new path only on a git-detected rename. A split
that leaves the source file alive (agent.rs, main.rs both still exist)
is not a rename, so the moved prose reads as newly written and is held
to a new file's zero-debt / grade-6.00 default. Fifteen of the
seventeen flagged constructions were reworded and dropped their debt
entirely (mostly bare issue-number citations, `docs/prose-guidelines.md`
rule 7). One could not be: `door.variant` on turn.rs:81 is a genuine
struct field access — the same field #5422 hit — unreducible without
renaming production code across ~24 call sites, well outside this PR's
scope. `scripts/prose-baseline.txt` carries one `rust-jargon` line for
it; `scripts/prose-grade-baseline.txt` carries turn.rs and
output_envelope.rs at their measured means (11.88, 10.93), since the
grade ratchet has the identical rename-only carry gap.

No witness: this is a pure code-organization move with no behavior
change, verified by re-reading the diff line-for-line and by
check-file-size.sh, check-prose.py, check-line-citations.py and
check-dead-code-allows.py all passing locally with no new baseline
debt beyond the two hand-edits explained above. CI's build and test
tiers are the ones this session cannot run locally.

Closes #5466
